+++ flowless-mvp-example/src/main/java/com/zhuinden/examplegithubclient/util/FlowlessActivity.java
@@ -26,7 +26,7 @@ public abstract class FlowlessActivity
                 .globalKey(getGlobalKey()) //
                 .dispatcher(this) //
                 .install(); //
-        transitionDispatcher.setBaseContext(newBase);
+        transitionDispatcher.setBaseContext(this);
         super.attachBaseContext(newBase);
     }