+++ flowless-sample-single-root/src/main/java/com/zhuinden/flowless_dispatcher_sample/MainActivity.java
@@ -29,7 +29,7 @@ public class MainActivity
                 .defaultKey(FirstKey.create()) //
                 .dispatcher(flowDispatcher) //
                 .install(); //
-        flowDispatcher.setBaseContext(newBase);
+        flowDispatcher.setBaseContext(this);
         super.attachBaseContext(newBase);
     }